requirements.txt
google-adk
litellm
mcp[cli]
mcp[server]
requests
requests-cache